Can now create a game history.
A text file will be created called records.txt, which will display the hand history of each player, for 1000 games.

History can be decoded as:
score of hand, score of hand + score of 3 community cards, score of hand + score of 4 community cards, score of hand + score of 5 community cards, score of 5 community cards

Might be an eror in holdem.py around line 221 about popping an empty list.  Not sure how this happens.